[Fix] Remove WOULD_LOG check from LOG_STREAM destructor
The LOG_STREAM macro is documented to NOT check the severity, yet it does.
With the introduction of tag-based log filtering this causes actual
issues as the application log tag will (mostly) be different to the one
used by the log library (which falls back to the binary name).
As the result log messages were wrongly ignored.

Change-Id: I5a9cbb853b5408d6ffd956bc14032271643dfe73